Optimize the inference of output paths
Summary:
After D39589411, I realized that all tito taint is stored under the same call info key in our taint representation.
This allows optimizing the `transform_non_leaves` function which is called whenever we read from a taint tree in the backward analysis, in order to add the output path in the return access path set. This operation can be optimized to be only performed on the call info for tito (i.e, `CallInfo.Declaration`). This should prevent iterations on all sink kinds.

Reviewed By: dkgi

Differential Revision: D39590895

fbshipit-source-id: ce06e2b55306af5424f64abcb015c5680f41dc0a