Make the call info explicit when creating taint
Summary:
This has bothered me for a while. The interface to create taint, `Taint.singleton`, takes an optional location and implicitly converts it to a `CallInfo.t`. This makes the interface confusing since the default `CallInfo.t` is implicit. It also clashes with our coding convention, we try to avoid optional parameters in practice.

This diff also makes it clear that all tito taint are always under `CallInfo.Declaration`, which can be used for some optimizations (see follow-ups).

Reviewed By: dkgi

Differential Revision: D39589411

fbshipit-source-id: 963ea04abe0525d76a923c273bd1c66eb779fb16