Implement the return access path set with the tree domain (1/n tito collapse depth)
Summary:
# Context
When applying taint-in-taint-out (i.e, tito) at a call site, we always collapse the input taint tree. This is both sound and allows us to scale pretty well, although it costs us precision and introduces false positives. Taint collapsing (also called broadening) means that we will consider the whole object as tainted if one or multiple attributes were tainted.

For instance:
```
def perfect_tito(x):
  return x

def issue():
  x = {"a": source(), "b": "safe"}
  y = perfect_tito(x) # only `x["a"]` is tainted, but collapsing leads to tainting `y` entirely.
  sink(y["b"]) # `y` tainted means `y["b"]` is also tainted, this triggers a false positive.
```

In this case, we could actually do better and infer that `perfect_tito` performs taint-in-taint-out without changing the structure of the taint, which means no collapsing is needed.

In the following diffs, we will implement the automatic inference of collapse depth, which is the depth at which we should collapse the input taint when applying tito. We currently collapse at depth 0 (i.e, turn the whole taint tree into a singleton) but could infer a maximum collapse depth (e.g, 4).

# This diff
In this diff, we implement the first step of inferring a collapse depth. We currently use a return access path set with a custom join. We replace this set with a tree domain which holds a collapse depth which will always be bottom or 0. Besides the change in the output format, this should be a non functional change.

Differential Revision: D39593946

fbshipit-source-id: 84611201489301fd11c5c1dd21df8175f9ec1a5d