Store the check-invariants option in the static analysis configuration
Summary:
In the previous diff, we added `--check-invariants` in the client, and set the `PYSA_CHECK_INVARIANTS` environment variable.
In this diff, we store the option in the static analysis configuration for consistency.
Note that this is non functional change for now, since the only way to use this in the binary is with the environment variable. This goal of this diff is just to have it available for later, if we want to add further checks.

Differential Revision: D39597355

fbshipit-source-id: ddf1352d2248529a6d87ae2e3ef527410ea3dd29