Codemod in a placeholder for file-level ocaml docs
Summary:
In order to make our codebase more accessible to new contributors, we
want all modules to have some (in most cases very short) documentation
about what the module does and, when it is not clear, why.

This commit putt placeholders in, we can grep the codebase for
`TODO(T132410158)` to start knocking them out.

Reviewed By: pradeep90

Differential Revision: D39662852

fbshipit-source-id: f92c40e365e487ae6f3670e94987dce81adc1df8