Store a mapping from sources/sinks to their matching sanitizers in source-sink filters
Summary:
The module `SourceSinkFilter` is responsible for keeping track of which sources can match which sinks. This is used when applying sanitizers to know if a source (resp. sink) can be removed, since it could never match with any sinks (resp. source).

The current implementation currently stores a `Sources.Set.t Sinks.Map.t`. This is inefficient since we need to translate `Sources.Set.t` into a `SanitizeTransformSet.t` in `should_keep_source` (resp. `should_keep_sink`). This diff adds another mapping that stores the matching sources/sinks as sanitize transforms. This should be better for performance.

This will also allow further performance optimization in a follow-up.

Differential Revision: D39677725

fbshipit-source-id: 0131e5237a5145f285e2484c0d070baadc366b90