(nfc] Reorder the sanitize transform module
Summary: I find it pretty confusing to have `SourceSet` and `SinkSet` defined on the top, where the whole point of the module is to define its `type t`. This diff moves the `type t` definition on the top and the `SourceSet`/`SinkSet` definition at the bottom. Also, moving `type t` closer to `module Source` and `module Sink` makes it easier to read and understand.

Differential Revision: D39679491

fbshipit-source-id: 7a280474853f007b2081c303fdb69b5f5bac061b