Break the cyclic dependency between taint transform operations and sources/sinks
Summary:
This diff is a non-functional change, but will be very important to allow a future optimization which relies on having `Sources.t` and `Sinks.t` available during taint transform operations.

This diff splits the definition on `TaintTransforms.t` and most operations on it in `TaintTransformOperation`. While it is less than ideal, this allows to use `Sources.t` and `Sinks.t` within `TaintTransformOperation` without creating a cyclic dependency.

Reviewed By: dkgi

Differential Revision: D39702725

fbshipit-source-id: 233281bcb2add728d4d4f48370bf65c9e60deca5