No Daemon Query: Create New Arguments type.
Summary: **Context**:  A quick factor that creates a new Arguments type to match the rest of the argument types in code (like check/analyze/infer, etc). I didn't do this originally because there were only 2 fields for queries, but I intend to add more such as a "sequential" flag, which is done in the next diff.

Reviewed By: stroxler

Differential Revision: D39578514

fbshipit-source-id: d5f109d0fbf0444388e6667d2f822a1bf1d4b47d