Add a new subscription variant to just get updates
Summary:
For a shadow-mode server, we want to disable type error handling.

I already implemented the logic not to pass type errors to the editor
on the client side, but it turns out the handling of the initial
`DisplayTypeError []` request when initializing a subscription crashes
a lazy-module-tracking server (because it tries to list all modules).
Moreover, any incremental update will also try to get all errors and
crash.

As a result, I need a way for servers with type checking disabled to
create subscriptions that explicitly ask for updates but not type errors.
I do this by making a new variant of subscription that asks for only
state updates but not type errors, and:
- using a Response.Ok for the initial request
- filtering subscriptions to just the ones that want type errors before
  lazily getting all type errors

Once I update the client to use the new subscription type whenever
the type_errors language server feature is disabled, this should allow
a lazy-module-tracked `pyre persistent` to work in shadow mode.

We can probably remove this code if we don't want to keep it once shadow
mode is using the new codenav daemon, this is just needed to unblock quick
iteration.

Reviewed By: grievejia

Differential Revision: D39586227

fbshipit-source-id: 6533e7d3013cbb4d3cdafda28c22244a1a8c2bda