Refactor subscriptions: create should take a request
Summary:
In order to unblock shadow mode, I need to add a new subscription
kind that just listens for state updates, but not type errors.

That's going to require Subscription.t to include some information
about the kind of subscription, and means that `create` needs to
know the actual request rather than just the name.

This refactor prepares for a world where there are multiple
subscription kinds and multiple request variants.

Reviewed By: grievejia

Differential Revision: D39586229

fbshipit-source-id: 10c9e25e80cc9c93239bf2ce8d3c1ad864353618