Add a feature to control status updates.
Summary:
For a shadow-mode server, we don't want status updates flowing
to vscode. We still do want them to be visible in python
so that we can see them in our logs and potentially use the status
in telemetry, but we don't want any user-visbile behavior, which could
be confusing.

Reviewed By: joshuaxu1995

Differential Revision: D39619345

fbshipit-source-id: 1a7b2c96f8c35145cf047904d92bfb82ddfc8888