Refactor subscriptions: add should just take a subscription
Summary:
Subscriptions contain names. Let's not pass another name in
separately, it makes the APIs more verbose and introduces
the possibility of a bug where we pass the wrong name.

I think maybe this API evolved out of a raw map add where
the name was a separate parameter but one of the benefits of
a custom API is we can use the fact that the name is already
there.

Reviewed By: joshuaxu1995, grievejia

Differential Revision: D39586230

fbshipit-source-id: bd660dd0ff6ca1ee1d26189c2c3762b2b930a16b