Structure user-facing docs for ease of use
Summary:
Noticed that it's extremely unclear what is or is not necessary for someone to read through to get a basic understanding of how to add type checking to the project.

Goal here is to:
- Make it more clear that installation/getting started page are redundant if someone is not trying to build from source or do something special
- Create a more sensible flow going from getting started to gradual typing information, perhaps skipping unnecessary details in installation or configuration

Will continue to adjust the contents of these pages a bit, but we want users to get all the information they need in one spot with Getting Started, and make it visually clear that all the other pages are optional "dive deeper" docs on a particular topic if they find it relevant.

Reviewed By: stroxler

Differential Revision: D39731793

fbshipit-source-id: 002f39002a725b77d1e7801fa85dc3fb9ad98dd3