Remove previous sanitize transforms from the local part when adding named transform
Summary:
As per the semantic of sanitizers, when the taint flows through a named transform, previous sanitizers are invalid (meaning, they do not apply anymore). This means it is safe to remove them from the local part of the taint transforms.

Note that this optimization cannot be done on the global part, since it is used to query the next frame in the UI. Any change would break the trace frame.

Differential Revision: D39704003

fbshipit-source-id: 5b236541402b93433c9f08a5ba36e03a57c3d99e