Add scalar breadcrumbs when reading from and writing into any int-typed variable
Summary:
Currently we only add breadcrumbs for function calls whose return type is
scalar, for performance reasons. At the cost of performance, we can
additionally add such breadcrumbs when reading from and writing into a
variable, because it is hopefully a cheap operation to query the type of a
variable (as opposed to an arbitrary expression, such as an object's
attribute).

Reviewed By: arthaud

Differential Revision: D39596825

fbshipit-source-id: 3d77d225f28729f78e8eff249f2f770e712be9e2