Remove ide overrides from top-level args
Summary:
This is the first commit of a stack to remove `ide_features.py` entirely;
it is no longer used at all because we give precedence to explicit
`persitent` command-line args and these are always set explicitly
by vscode.

It removes the no-longer used flags from command-line entrypoints.

The stack, which I'm doing in a few stages to keep things easy to
review, will have at least two more commits:
- removing the IdeFeatures class from the configuration
- removing the fields from CommandArguments

Reviewed By: grievejia

Differential Revision: D39891070

fbshipit-source-id: 0f2b8fcfc795aa4deec9ab7395972adf419d5dbd