Remove `ide_features.py`
Summary:
This config does nothing now that we tell `pyre persistent` which
features we actually want. Getting rid of it removes a surprising
amount of boilerplate code, especially tests.

Note that LanguageServerFeatures.create is now in a slightly weird state where
it accepts a `configuration` argument it does not use and the various feature
flags don't come in a consistent way. I'll clean this up in a few commits, but
to keep diffs small I didn't want to introduce unnecessary changes to
persistent-specific type signatures in this commit.

Reviewed By: grievejia

Differential Revision: D39891564

fbshipit-source-id: 603d48bf6e4bb6c05d1a2e174d3c933aecb58f1c