Remove dead fields from CommandArguments
Summary:
Now that we are no longer constructing an IdeFeatures instance from
CommandArguments, we can get rid of the ide feature flags.

This completes the removal of `ide_features.py` and directly related
logic, although I will follow up with a clean up diff that simplifies
the logic for `pyre persitent`, which was complicated because it had
to be forward/backward compatible and can now be made much more
straightforward.

Reviewed By: grievejia

Differential Revision: D39891677

fbshipit-source-id: ab62ca76c9fe4ef9e25d947fae2446e32ff7ef7a