Import language_server_features module
Summary:
We decided that for new code, we want to avoid importing function and class
names directly and instead import module names.

But `language_server_features` is a module, and importing it directly saves
quite a few super-long lines that were a little hard to read (and were pep-8
violaions, a couple lines were well over 90 characters but there was no way to
autoformat them because the names were so long).

Reviewed By: grievejia

Differential Revision: D39894045

fbshipit-source-id: 5d27fb07851b7f7b2dd1045bb9fdc0966dcc233b