Silence missing-method errors: add test harness
Summary:
I'm trying to get rid of the spam we get in our logs that looks like this:
```
DEBUG Exception occurred while reading JSON RPC: Required field `method` is missing: {'jsonrpc': '2.0', 'id': 0, 'result': None}
```

The problem here is that vscode sends us empty pings that don't have a method,
which we should just jump over but we currently treat it as an error and log
exceptions, which junk up the logs and make development more difficult.

Adding a pair of tests of the current behavior, so that it will be clearer how
my fix works in the next commit.

Reviewed By: grievejia

Differential Revision: D39904002

fbshipit-source-id: 0de3a613aec0ad55ca87412a59bdcca617ff7508