Fix shadow-mode bug: response to state change subscription is ["Ok"]
Summary:
Prior to this commit, the `_subscribe` code for a shadow-mode language
server with type errors turned off was sending a "SubscribeToStateChanges"
request to the daemon but then trying to interpret the response - which
is suppused to be `["Ok"]\n`, as a list of type errors.

This caused the background task to crash, which actually doesn't bock
initial launch of shadow mode because we don't need the background task
except for uptime metrics, but it's obviously bad and also makes it
*very* hard to look at the logs and be confident things are okay.

The solution is to separate the full subscription logic, in particular
the handling of the initial response.

Reviewed By: grievejia

Differential Revision: D39952767

fbshipit-source-id: 5ec70d27f06eea2312c8501177b5d48d511d2aaf