Add duration_ms to telemetry events
Summary:
As I was testing out shadow-mode, I noticed that the chronicle events
don't actually include duration for telemetry. This makes a kind of sense;
I was expecting chronicle to compute duration since start of the activity,
but really telemetry is point-in-time data so that wouldn't necessarily
be meaningful.

As a result, I'm adding duration computed in python to the telemetry dict.

Reviewed By: joshuaxu1995

Differential Revision: D39975899

fbshipit-source-id: 21d49c18f4edd80ea362f208acb9a5ab6a537d8a