Correctly look up same module global taint
Summary:
Before this change, Pysa does not correctly look up models from global variables defined and used in the same module:
```
# .pysa
foo: TaintSource = ...

#.py
foo = []
def f():
    _test_sink(foo)

#.models
<f not shown in issues>
```

The reason is that in identifier look up, the forward analysis always assumes that the model should be looked up in the current scope. However, global models are not stored in the current scope.

This diff attempts to fix the issue by combining models in the GlobalModel and the current scope for global variable model look ups.

Reviewed By: arthaud

Differential Revision: D39866810

fbshipit-source-id: d5caf849b693969a4adeefcd2d07a97cfec8d764