Log state of server with every change request
Summary:
**TLDR**: This diff attempts to publish telemetry events whenever a didChange event happens, along with the state of the server, when that event happens.

**Context**: This is part of a diff stack that tracks server uptime. See this doc for more info on the overall plan.

We want to sample `didChange` events and figure out the state of the server during those events. That serves as a much more accurate proxy for the amount of time the server is actually up, because when a user actually edits a file is when we care most that the server is ready.

Reviewed By: sinancepel

Differential Revision: D39521918

fbshipit-source-id: cfe3c775ef18c27a63aa9e128cc8ade1639bc2d7