Demonstrate a false negative due to inferring the wrong output path
Summary: We use the tree domain for return access paths, which enforces a downward closure in the tree, i.e if a node has an element, then inserting another element that is smaller or equal in any of its children is a no-op. This leads to false negatives for return access paths (i.e, output path of tito), as demonstrated by this diff.

Reviewed By: sinancepel

Differential Revision: D40080653

fbshipit-source-id: 4287f82f9dc45121ec2fac291f5770838f2441a9