Add a new sink to mark the endpoints of the extra traces and Create ExtraTraceSink when introducing taint transforms
Summary:
A frame is a call edge from one function to another. This stack of diffs will
add subtraces to frames to show taint transforms that propagate the taint
within the frame, i.e., taint transforms that are called directly or indirectly
by the callee in a frame.

A subtrace is implemented as a normal sink trace that ends with a (special) sink.
It can be used to show tito paths (i.e., sequences of function calls that
propagates the taint), which may contain taint transforms.

Reviewed By: arthaud

Differential Revision: D39901562

fbshipit-source-id: 845479fe17f96a93c3803e21d6428f6ad0186862