Overlays: Use base environment instead of skipping query if overlay id is invalid
Summary:
**TLDR**: For overlays we are dropping all queries where the overlay id is incorrect, but it might be better to fall back to the base environment in these case.

**Context**: For any query handled by the server, there is an optional additional field called overlay_id (which is equivalent to the file name). If the overlay_id is invalid, we do not process the query. However, there are cases in which an overlay_id is guaranteed to be invalid.

Currently, the overlay id would be invalid in cases in which the client makes a query on a file that hasn't been modified - this is because the action that will create the overlay environment is the first update of a file. Therefore, if the file hasn't been touched, then the no overlay environment will exist.

NOTE: The use case where this would be a problem would be the first time a user runs a go-to-def request without modifying a file first. The user would expect that the go-to-def request would work properly- but it does not because the overlay_id for that file does not exist, and the resulting behavior is an error when it should be a valid go-to-def response.

In these cases, we want to fallback to the base environment and log to the server that the query contained an invalid overlayId. This will give the user a response based off the saved file rather than the overlay, which seems like reasonable behavior (for now).

**Alternatives Considered**:
1. I considered an alternative of creating an overlay if one doesn't exist, then use that environment. However, I decided against this approach because overlay environments can be pretty expensive and we don't want to store an environment if it is the exact same as the base env.

2. I also considered storing a flag on the clientside to denote whether a file has been modified before and if it hasn't, pass a null overlay id instead of the actual overlay_id (file_name). But I thought this solution might invalid storing some unnecessary state and is not very elegant.

Reviewed By: grievejia

Differential Revision: D40114627

fbshipit-source-id: ee4e6be333d2e0a50dda5f81ca77170c959edbc4