Pass in type resolution in the readonly fixpoint.
Summary:
Problem: We want to use the output of the type checking analysis to do things like get the callable type of a function, get the overload that was actually selected for that function call, etc.

This diff just stores the local annotations (from the type checking analysis) in the context and uses that to pass down the type resolution to the `forward_xxx` functions.

There is no behavior change yet. That will come in the later diffs.

Reviewed By: sinancepel

Differential Revision: D40183478

fbshipit-source-id: 043019e5cc661b5a6e88b52b46b543e503ced650