Remove logic to set `__override__` in `@override`
Summary:
Including this code and the tests was useful for driving discussion
of PEP 698, but eventually we got 100% agreement that it is better
not to have any runtime behavior for `typing.override`.

As a result, we do not need runtime behavior in pyre extensions either.

I left the test classes in place because now that we have them it is
maybe nice to leave an example, but I would also be happy to remove it
entirely.

Reviewed By: sinancepel

Differential Revision: D40251091

fbshipit-source-id: ea83c4a2d5c5681ada3fbea0b4ee9899c49f8f9f