Log LSP Request time for expression-level coverage
Summary:
Before rolling out expression level coverage to 100%, I'd like to see the p95 for the LSP events. This will help us determine if we need to optimize it further before rolling it out. Ideally, it's opt-in for the squiggles (with the button in the gutter: {F780540826} ) but the [gk check](https://www.internalfb.com/intern/gatekeeper/projects/pyre_vscode_expression_coverage/) is on by default.

There is already a `{"commandName":"pyre.toggleTypeCoverage"}` with the duration of the check but it doesn't actually log unless someone enables expression level coverage from the gutter. This logging (each time there is a typecoverage check - even if button in gutter is disabled) should help us determine if it'll slow down the typechecking server by enabling the GK. "shadow mode" you can call this.

stroxler do you know if that's something VSCode gives us by default? and is this diff a good approach? is it easy to add this to a dashboard?

alternate approaches:
decorator: D40313828 - I thought this would be great to reuse for all expressions but even with how generic it is it isn't able to log differently based on result of LSP request - we might get weird outliers low times and not know that the request actually failed. Sinan suggested a contextmanager but I think either of these options are overkill for this (for now).

Reviewed By: stroxler

Differential Revision: D40315589

fbshipit-source-id: 6b10172bd7fa3865e12fabe17aa04310a4546785