Treat constants as mutable.
Summary:
Problem: The diff title is oxymoronic, I know. We were treating `42` as `ReadOnly`, which seems reasonable, but leads to an error when writing `x: int = 42`, which is not reasonable.

If were to overengineer this, we would treat constants as generic function calls `x: int = make_const<T>()`, where we need to use the expected type to solve for `T` (i.e., using bidirectional inference). I don't think we want to go that route.

I think it's much simpler to treat them as `Mutable`, since they are assignable to both `Mutable` and `ReadOnly`.

Reviewed By: stroxler

Differential Revision: D40315043

fbshipit-source-id: a442770a54cc9dcec836dd330c43548c3951cd51