Add basic function to get callable data from a callee.
Summary:
Problem: Given a function call `foo(x, y)`, we want to get the callable type of `foo`, so that we can check the readonlyness of each argument against its parameter.

This diff just adds a function that does the above for simple, non-overloaded callable types. There is only possible signature here, so it's straightforward.

There are other cases, such as methods, overloads, unions, `TypeVar`s, etc. For these, we need to know the overload selected by the type checking analysis. Will handle those later in the stack. Added a TODO comment for now.

Next diff: Use the output of this function to check the readonlyness of arguments against parameters.

Reviewed By: sinancepel

Differential Revision: D40266726

fbshipit-source-id: 646731b50356ea9c7dd8f205ed6ec42b802a91f4