Disable tito collapsing entirely for named tuples and dataclasses
Summary: For named tuples and dataclasses, we know that the constructor is trivial (i.e, arguments are propagated to their attributes). We do not need to collapse on tito for those. Instead of using the maximum collapse depth from the taint configuration (currently set to 4), this diff uses `max_int-1`, to entirely disable collapsing. Since this is done for few functions, this should (hopefully) be fine in terms of performance. This will also be used to implement a `NoCollapse` annotation in the future.

Reviewed By: alexkassil

Differential Revision: D40287428

fbshipit-source-id: daff14edc746bb004c8adb24e6ce8604ade575e2