Rename `path` into `applies_to`, in preparation for adding `ParameterPath` and `ReturnPath`
Summary:
Previous context: D39593946 (https://github.com/facebook/pyre-check/commit/5ef990d11ac834c104f8fb92be1164137a18f1dc)

In the following stack, we will add ways for user to specify input and output paths for tito, as well as collapse depth.
There is already a way to specify the input path using `AppliesTo`, which we will deprecate in the future.
To prepare for the transition, this diff explicitly renames `path` into `applies_to` in the structure that holds the parsed value for a source, sink or tito.

Reviewed By: alexkassil

Differential Revision: D40403683

fbshipit-source-id: b1a951fa41d29023640f5f52856cfc0e255e98a5