Return Result types for Failed Queries rather than throwing exceptions
Summary:
**TLDR**: Instead of propagating exceptions due to failed queries all the way up the call stack to be caught, instead return Result types to be processed by the caller.

The reason we want to return result types is that we already know in advance that it's a very possible circumstance that we will encounter a query error. Borrowing from a principle in functional programming, we are adapting an either type pattern where we will return either a valid response, or a Failure. In case you have no context on this topic, here's a helpful article (https://dev.to/anthonyjoeseph/either-vs-exception-handling-3jmg)

The purpose of this diff stack is to log exceptions for LSP events to scuba. Before that can be done, there are issues in which I cannot log exceptions to scuba because they are being caught at too low a level to be written to telemetry. This diff focuses on getting us to a state where we can log those exceptions, and in a future diff, we will actually perform that logging.

I changed the unit tests to reflect that most of the operations throw exceptions now rather than returning empty responses.

Reviewed By: stroxler

Differential Revision: D40239217

fbshipit-source-id: 8bb70c4d08c7e7cf520810d681ad564bbd517da6