Add logging for scuba events for exception tracing.
Summary:
**TLDR**: This diff logs the exception string to scuba to help understand the causes of why a specific user event fails.

The previous diffs changed the code structure of exception handling to use the functional programming paradigm of `Result` rather than using exceptions, followed by propagating the result up to the layer where it can be logged. This diff takes care of the actual step of logging the error string to scuba.

Reviewed By: stroxler

Differential Revision: D40385655

fbshipit-source-id: a95158c2627e9c638696a1f7abf4b437f63e41f9