Add extension tests to the module tracker suite
Summary:
We have confirmed that Pyre's lazy module tracker has a bug when
dealing with custom extensions.

Before we attempt to fix the problem, I would like to record the bug in
tests so that we can see the fix take effect.

What we see in this new unit test is that lazy tracking handles a
"standard" custom extension okay, but when the custom extension
suffix is supposed to be included in the qualifier we have a bug.

Reviewed By: kinto0

Differential Revision: D40456515

fbshipit-source-id: b394f63c3c54dfb4b6da7e0250dd4de429551c3d