Implement parsing of access paths for pysa models
Summary:
In the following stack, we will allow user models to specify complex access paths for taint-in-taint-out, as well as the collapse depth. This will reduce false positives by better modeling the actual behavior of functions and methods.

This diff implements the parsing of access paths. Access paths starts with an underscore `_` which represents the current argument or return value. The underscore can be followed by item accesses `_["a"]["b"]` or attribute accesses `_.a.b`, or a mix of both. Note that the current analysis never uses attribute accesses, so in most cases, you want to use item accesses. Access paths can also use `.keys()` which represents all keys of a dictionary. Finally, `.all()` represents all possible indexes.

For instance, one could model the taint-in-taint-out for the given function:
```
def foo(arg, i):
  return {"foo": arg.attribute[i].keys()}
```
Using the given model:
```
def foo(arg: TaintInTaintOut[ParameterPath[_["attribute"].all().keys()], ReturnPath[_["foo"]]]): ...
```
Note how `arg.attribute` is represented as `_["attribute"]` because the analysis does not differentiate attributes and dictionary accesses.
Also note that `.all()` is used because we do not know the value of `i` at analysis time.
This diff only implements the parsing of access paths. `ParameterPath` and `ReturnPath` will be added in a follow-up.

Differential Revision: D40447221

fbshipit-source-id: 0286a40812d0e8579612dc2673b42c9d9ee0906d