Start pyre daemon for go to def request if daemon is down
Summary:
**TLDR**: We need to restart the pyre daemon if it is down, for the go to def event.

Part of a larger diff stack to measure vscode downtime causes: D39521918 (https://github.com/facebook/pyre-check/commit/5464577e130dc6b23b784a6f59f428cb427861ef). The most significant event we care about is go-to-def.

The scuba logging that motivated this diff was https://fburl.com/scuba/vscode_chronicle_language_server_events/14xhrl32. You can see that there are a long set of consecutive go to def requests that all have a server state of disconnected- it would be better to restart the pyre server in these cases so that the daemon is available for go-to-def in that case. Additionally, you can see here (https://fburl.com/daiquery/rk6qm1lu) that the requests with 'definition' from date 10-17-22 had a fairly large % that occurred when the server was not connected - 47.3% (79/167) of go to def events occurred when the daemon was disconnected.

This diff will restart the pyre daemon if it is down during a `definition` event. The server uptime should be improved because if we ever detect the pyre server is down on a go to def request, we will restart it.

**Note**: I thought about doing the same for find reference and hover but decided against it. For hover, the event might occur too frequently. For find references, we're not currently making a call to the pyre daemon so it doesn't make sense to restart it there either.

Reviewed By: grievejia

Differential Revision: D40184715

fbshipit-source-id: c0b34b099a24433040393e6199400b8df0c33237