support `__qualname__`'s for all callables
Summary:
As per [PEP 3155](https://peps.python.org/pep-3155/), all functions and classes have associated
qualnames. A callable is either a function or an object with a `__call__` method, so this diff adds
the `__qualname__` attribute to Pyre's understanding of callables, which'll cover one part of the
PEP (types are already covered properly via the builtins typeshed stubs).

Note that this change isn't completely accurate - see a test where Pyre doesn't error where it
should. The root cause of that issue has to do with Pyre's limited understanding of callables.
Technically, Pyre should be viewing functions as belonging to the `function` subclass of callables,
but that would require a major re-jiggling of the type architecture. I chose to cover the
overwhelmingly common case in this diff with the risk of missing some type errors.

Reviewed By: stroxler

Differential Revision: D40512674

fbshipit-source-id: 2395b68c3eefa11e0fd1642aee61c396ec2d8857