Emit error when assigning to readonly attribute.
Summary:
Problem: We want to emit an error in two cases: (a) the base is readonly (b) the base is mutable but the attribute is explicitly readonly.

Note: We don't error on readonly variables being assigned again:

```
x: ReadOnly[int] = 42
x = 99
```

Since this isn't strictly mutating an existing readonly object; it's reassigning to the variable. RFC if people have strong opinions otherwise.

Reviewed By: stroxler

Differential Revision: D40557393

fbshipit-source-id: 7a10656f0811ca15fb81969d04b7a1b899836d3f