Add `__override__` back to `@override`
Summary:
A few weeks ago I took out the setting of the `__override__`
field in `override` because after a first round of PEP 698
discussion we tentatively decided against it.

After feedback from the author of the `overrides` library as
well as revisiting the spec and implementation of `typing.final`
(which specifies no runtime behavior, but in fact setting a
`__final__` attribute was added later because we realized that
it is helpful to have) we decided to pivot back to including the
this logic.

The implementation here is lifted directly from `typing_extensions` as of
https://github.com/python/typing_extensions/commit/6e72c09ff6b501aaf5abed601afbcd1986ae0bd2
and is more robust than the initial implementation - for some special
descriptor classes the `setattr` attempt could fail so usign a try/except
is a good idea.

The tests are also lifted directly from `typing_extensions`, other than I
removed a not-so-important assert that depends on utilities we don't have
and I added type annotations.

Reviewed By: sinancepel

Differential Revision: D40569325

fbshipit-source-id: 1713fd89f70dd7abf0daaee34f3113e505189e08