Get server status before and after request is handled.
Summary:
We have a problem right now with how we're implementing server
statuses. We're getting the status after the request is complete rather than before the request is handled, which likely paints a rosier picture of the server uptime. This fix will log the server status before a request comes in.

**Question: Why log 2 states**?

This diff logs the status before and after a request. It's important to keep the status AFTER the request because it's possible that when the request gets handled, the pyre server goes down. When I was testing a change where the status was being logged before the request in V1, I observed there were case in which the server was in a ready state prior to the request handling, then the server went down during the request processing, which caused a false negative where we didn't identify a disconnected state where we should have.

As a result, we would have two columns that both need to display "READY" in order for us to validate that the pyre server was truly ready.

I realize that this presents more complicated cases in the data pipeline processing. However, I feel that two columns would ultimately provide more accuracy for us to evaluate uptime during definition requests.

Reviewed By: stroxler

Differential Revision: D40588767

fbshipit-source-id: f90e617914ca8093830688b210eb04234d0de008