Remove dependency on `TypeCheck`.
Summary:
Problem: We want to call the readonly analysis after the type check analysis. However, this causes a circular dependency, because the readonly analysis unnecessarily uses `TypeCheck`.

So, inject just the dependencies we care about - namely, the global resolution, the local annotations, and the type resolution for each statement.

Next diffs:
+ Plug it at the end of type check analysis.
+ Add a config flag in the client to enable readonly analysis

Reviewed By: sinancepel

Differential Revision: D40593613

fbshipit-source-id: 115eff7615f956d09cde5a9989804c1c942aacf8