Key module by entire qualifier in cache
Summary:
Lazy module tracking searches for x.y.z at x/y/z. Configerator imports are of the form `x.y.z.cinc`. By that logic, we'd be looking in `x/y/z/cinc`.

To support these extensions, we look in the config for this bit:
https://pxl.cl/2hsCn

If it's present, we allow importing with that suffix and treat `z.cinc` as the qualifier. We already support this in the classic module finder and thought this (D40456457 (https://github.com/facebook/pyre-check/commit/57f1fc5f6dd779c634d6707fbdf22832e9f4d15c)) fixed it for lazy but there's still one problem: the cache is keyed by `last` in the reference (the `cconf` part).

This changes the key to be the entire qualifier reference.

Reviewed By: stroxler

Differential Revision: D40568549

fbshipit-source-id: e0b3506928aee0a39e173144eaa8005dc7a0878c