write a SocketCreated event to stdout when socket's created
Summary:
The codenav client will need to pick up on a socket being created. The modern (tm) way of ensuring we don't have race conditions is to have the server process signal via stdout when the socket's ready. We already do so on the new server but not on code nav, so let's add it here!

Just like the regular newserver, the events go on stdout as part of the initialization protocol. A minor design decision I made was to mark the server as listening after the full init instead of adding another hook. Since everything is lazy this should be fast enough and not matter, but it is worth noting.

Reviewed By: stroxler

Differential Revision: D40658747

fbshipit-source-id: 050be9b1982240bfc17753978e39bbf4b4f3e222